Add files via upload
Note that this is a RAR file. You must use unrar on Linux/Mac or WinRar on windows to properly unpack this file otherwise it will appear corrupt. This file is the full data set and is ~400 MB uncompressed. It is not needed for the scripts as they use the wlinn_sample file